Support different Executors. (#90)
