+++ src/lib/tests/PGManagerTest.cpp
@@ -1,7 +1,4 @@
 #include <boost/uuid/random_generator.hpp>
-#include <boost/uuid/uuid_io.hpp>
-#include <folly/init/Init.h>
-#include <folly/executors/GlobalExecutor.h>
 
 #include <homeobject/pg_manager.hpp>
 #include "lib/tests/fixture_app.hpp"
@@ -57,15 +54,4 @@ TEST_F(TestFixture, Migrate) {
               PGError::INVALID_ARG);
     EXPECT_FALSE(
         homeobj_->pg_manager()->replace_member(_pg_id, _peer2, PGMember{boost::uuids::random_generator()()}).get());
-}
-
-int main(int argc, char* argv[]) {
-    int parsed_argc = argc;
-    ::testing::InitGoogleTest(&parsed_argc, argv);
-    SISL_OPTIONS_LOAD(parsed_argc, argv, logging);
-    sisl::logging::SetLogger(std::string(argv[0]));
-    spdlog::set_pattern("[%D %T.%e] [%n] [%^%l%$] [%t] %v");
-    parsed_argc = 1;
-    auto f = ::folly::Init(&parsed_argc, &argv, true);
-    return RUN_ALL_TESTS();
 }
