+++ test_package/test_package.cpp
@@ -1,10 +1,11 @@
+#include <folly/init/Init.h>
 #include <sisl/options/options.h>
 #include <homeobject/homeobject.hpp>
 #include <boost/uuid/random_generator.hpp>
 
 SISL_LOGGING_INIT(HOMEOBJECT_LOG_MODS)
 
-SISL_OPTIONS_ENABLE(logging)
+SISL_OPTIONS_ENABLE(logging, homeobject)
 
 class TestApp : public homeobject::HomeObjectApplication {
 public:
@@ -18,9 +19,12 @@ public:
 };
 
 int main(int argc, char** argv) {
-    SISL_OPTIONS_LOAD(argc, argv, logging)
+    SISL_OPTIONS_LOAD(argc, argv, logging, homeobject)
     sisl::logging::SetLogger(std::string(argv[0]));
     spdlog::set_pattern("[%D %T%z] [%^%l%$] [%n] [%t] %v");
+
+    auto parsed_argc = 1;
+    auto f = ::folly::Init(&parsed_argc, &argv, true);
 
     auto a = std::make_shared< TestApp >();
     homeobject::init_homeobject(a);
